restructured developer metadata slightly
